fix - refactor directory structure. Passed method in the context. add - documentation
