+++ src/main.ts
@@ -5,7 +5,7 @@ import { Schema } from "./schema";
 import * as cors from "cors";
 import * as helmet from "helmet";
 import * as morgan from "morgan";
-import { persons } from "./schema/data-base/person-database";
+import { persons, findPerson, addPerson } from "./data-base/person-database";
 
 // Default port or given one.
 export const GRAPHQL_ROUTE = "/graphql";
@@ -37,7 +37,9 @@ export function main(options: IMainOptions) {
 
     app.use(GRAPHQL_ROUTE, bodyParser.json(), graphqlExpress({
         context: {
-            persons
+            persons,
+            findPerson,
+            addPerson
         },
         schema: Schema,
     }));
