+++ src/schema/modules/mutation.ts
@@ -1,4 +1,4 @@
-import { addPerson } from "../data-base/person-database";
+import { addPerson } from "../../data-base/person-database";
 
 export const typeDef = `
 # Mutations
@@ -10,7 +10,7 @@ type Mutation {
 export const resolver = {
     Mutation: {
         addPerson(root, args, ctx) {
-            return addPerson(ctx.persons, {id:  Math.random().toString(16).substr(2), name: args.name, sex: args.sex});
+            return ctx.addPerson(ctx.persons, {id:  Math.random().toString(16).substr(2), name: args.name, sex: args.sex});
         },
     },
 };
